Switch to single-precision parameters.

+++ src/ngraph/op/fused/hard_sigmoid.cpp
@@ -27,7 +27,7 @@
 using namespace std;
 using namespace ngraph;
 
-op::HardSigmoid::HardSigmoid(const shared_ptr<Node>& data, double alpha, double beta)
+op::HardSigmoid::HardSigmoid(const shared_ptr<Node>& data, float alpha, float beta)
     : FusedOp("HardSigmoid", {data})
     , m_alpha(alpha)
     , m_beta(beta)
@@ -41,17 +41,17 @@ NodeVector op::HardSigmoid::decompose_op() const
     auto data_shape = data->get_shape();
     size_t elem_count = shape_size(data_shape);
 
-    std::shared_ptr<ngraph::Node> alpha_node = ngraph::op::Constant::create<double>(
-        data->get_element_type(), data_shape, std::vector<double>(elem_count, m_alpha));
+    std::shared_ptr<ngraph::Node> alpha_node = ngraph::op::Constant::create<float>(
+        data->get_element_type(), data_shape, std::vector<float>(elem_count, m_alpha));
 
-    std::shared_ptr<ngraph::Node> beta_node = ngraph::op::Constant::create<double>(
-        data->get_element_type(), data_shape, std::vector<double>(elem_count, m_beta));
+    std::shared_ptr<ngraph::Node> beta_node = ngraph::op::Constant::create<float>(
+        data->get_element_type(), data_shape, std::vector<float>(elem_count, m_beta));
 
-    std::shared_ptr<ngraph::Node> one_node = ngraph::op::Constant::create<double>(
-        data->get_element_type(), data_shape, std::vector<double>(elem_count, 1.0));
+    std::shared_ptr<ngraph::Node> one_node = ngraph::op::Constant::create<float>(
+        data->get_element_type(), data_shape, std::vector<float>(elem_count, 1.0));
 
-    std::shared_ptr<ngraph::Node> zero_node = ngraph::op::Constant::create<double>(
-        data->get_element_type(), data_shape, std::vector<double>(elem_count, 0.0));
+    std::shared_ptr<ngraph::Node> zero_node = ngraph::op::Constant::create<float>(
+        data->get_element_type(), data_shape, std::vector<float>(elem_count, 0.0));
 
     return {std::make_shared<op::Minimum>(
         std::make_shared<op::Maximum>(alpha_node * data + beta_node, zero_node), one_node)};
